requirements.txt
# NovaOS Requirements
# Generated: 2025-12-06
# Python 3.10+ required

# =============================================================================
# CORE DEPENDENCIES
# =============================================================================

# OpenAI API client - used by LLMClient for chat completions
# Uses new SDK pattern: from openai import OpenAI
# Minimum 1.0.0 for new client API, pinned to latest stable
openai>=1.0.0,<3.0.0

# Flask web framework - used by nova_api.py for REST endpoints
# Provides /nova POST endpoint for web/mobile clients
flask>=3.0.0,<4.0.0

# Environment variable loading - loads .env for API keys
# Used in llm_client.py: from dotenv import load_dotenv
python-dotenv>=1.0.0,<2.0.0

# =============================================================================
# MEMORY SYSTEM (v0.11.0)
# =============================================================================

# Sentence Transformers - used for semantic memory search
# Provides embedding-based retrieval for better memory matching
# Falls back gracefully to keyword search if not installed
sentence-transformers>=2.2.0,<4.0.0

# =============================================================================
# NOVA COUNCIL (Multi-Model Orchestration)
# =============================================================================

# Google Generative AI SDK - used for Gemini Flash/Pro integration
# Provides cheap quest ideation (Flash) and research (Pro) capabilities
# Falls back gracefully to GPT-5 only if not configured
google-generativeai>=0.8.0,<1.0.0

# =============================================================================
# REMINDERS SYSTEM (v2.0.0)
# =============================================================================

# Timezone database for zoneinfo module
# Required on Windows; provides IANA timezone data for reminder scheduling
# Linux systems typically have this from the OS, but including for portability
tzdata>=2024.1

# =============================================================================
# OPTIONAL / DEV DEPENDENCIES (uncomment if needed)
# =============================================================================

# Rich terminal output (optional - for better CLI formatting)
# rich>=13.0.0

# Colorama for cross-platform terminal colors (optional)
# colorama>=0.4.6

# Pydantic for data validation (optional - if adding strict typing)
# pydantic>=2.0.0

# =============================================================================
# NOTES
# =============================================================================
# tkinter is part of Python stdlib (comes with Python installation)
# No additional install needed for: json, os, re, datetime, pathlib, etc.
#
# For development/testing, consider also installing:
#   pytest>=7.0.0
#   black>=23.0.0
#   flake8>=6.0.0
